Fixed a small error in one_d_poisson_hp_adapt.cc so that the test passes (again)
Added an additional output function to 
navier_stokes_surface_drag_torque_elements.h

Modified the source so that it will compiler using the IBM xlc compiler under
AIX. This required changing a few iterators to const_iterators, moving
one destructor into the .cc file rather than the .h file and "casting" the 
arguments of certain mathematical functions to be double rather than integer.
Suitable config files are provided in configure_options.

Build machinery that still needs to be modified is a specific option to 
include non-optimised fortran compiler flags and to include the stacktrace
library explicitly.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@276 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ src/generic/nodes.h
@@ -1071,7 +1071,7 @@ public:
   for (int i=-1;i<nval;i++)
    {
     hang_code+=unsigned(Node::is_hanging(i))*
-     unsigned(pow(unsigned(2),unsigned(i+1)));
+     unsigned(std::pow(2.0,double(i+1)));
    }
   return hang_code;
  }
